refactor: MainHeader를 공용 컴포넌트로 이동하고 AppHeader로 리네이밍 (#78)
