+++ frontend/src/features/main/components/index.ts
@@ -1,2 +1,1 @@
 export { default as MainContent } from './MainContent';
-export { default as MainHeader } from './MainHeader';
